Trait connecteer_translation::buffer::Buffer

pub unsafe trait Buffer {
    type Reader<'a>: Read<Error = Infallible> + 'a
       where Self: 'a;

    // Required methods
    fn feed_bytes(&mut self, bytes: &[u8]) -> usize;
    fn get_read(&mut self) -> Self::Reader<'_>;
    fn discard_read_bytes(&mut self);
    fn keep_read_bytes(&mut self);
}

Required Associated Types

type Reader<'a>: Read<Error = Infallible> + 'a where Self: 'a

Required Methods

fn feed_bytes(&mut self, bytes: &[u8]) -> usize

Feed bytes into the Buffer The return value is the number of bytes that were copied into the buffer, so the caller knows which bytes to keep (if any)

fn get_read(&mut self) -> Self::Reader<'_>

Get a Reader into the Buffer This reader needs to NOT discard bytes while reading since there are dedicated methods for discarding these bytes (or resetting read head)

fn discard_read_bytes(&mut self)

Discard every bytes that were read from an Self::Reader

fn keep_read_bytes(&mut self)

Keep the bytes that were read into the buffer, allowing them to be re-read when a new Reader is reading into the buffer